[metadata manager m8s] [Validator] comment resolution
